ai_service/requirements.txt
flask==3.1.0
flask-cors==5.0.0
numpy==2.2.1
scikit-learn==1.6.1
scipy==1.15.1
google-generativeai==0.8.3
python-dotenv==1.0.1
